+++ repos/cache/src/commonMain/kotlin/dev/inmo/micro_utils/repos/cache/fallback/crud/AutoRecacheCRUDRepo.kt
@@ -1,0 +1,36 @@
+package dev.inmo.micro_utils.repos.cache.fallback.crud
+
+import dev.inmo.micro_utils.repos.CRUDRepo
+import dev.inmo.micro_utils.repos.WriteCRUDRepo
+import dev.inmo.micro_utils.repos.cache.cache.FullKVCache
+import dev.inmo.micro_utils.repos.cache.fallback.ActionWrapper
+import kotlinx.coroutines.CoroutineScope
+import kotlin.time.Duration.Companion.seconds
+
+open class AutoRecacheCRUDRepo<RegisteredObject, Id, InputObject>(
+    originalRepo: CRUDRepo<RegisteredObject, Id, InputObject>,
+    scope: CoroutineScope,
+    kvCache: FullKVCache<Id, RegisteredObject> = FullKVCache(),
+    recacheDelay: Long = 60.seconds.inWholeMilliseconds,
+    actionWrapper: ActionWrapper = ActionWrapper.Direct,
+    idGetter: (RegisteredObject) -> Id
+) : AutoRecacheReadCRUDRepo<RegisteredObject, Id>(
+    originalRepo,
+    scope,
+    kvCache,
+    recacheDelay,
+    actionWrapper,
+    idGetter
+),
+    WriteCRUDRepo<RegisteredObject, Id, InputObject> by AutoRecacheWriteCRUDRepo(originalRepo, scope, kvCache, idGetter),
+    CRUDRepo<RegisteredObject, Id, InputObject> {
+
+    constructor(
+        originalRepo: CRUDRepo<RegisteredObject, Id, InputObject>,
+        scope: CoroutineScope,
+        originalCallTimeoutMillis: Long,
+        kvCache: FullKVCache<Id, RegisteredObject> = FullKVCache(),
+        recacheDelay: Long = 60.seconds.inWholeMilliseconds,
+        idGetter: (RegisteredObject) -> Id
+    ) : this(originalRepo, scope, kvCache, recacheDelay, ActionWrapper.Timeouted(originalCallTimeoutMillis), idGetter)
+}