+++ repos/cache/src/commonMain/kotlin/dev/inmo/micro_utils/repos/cache/util/ActualizeAll.kt
@@ -2,60 +2,52 @@ package dev.inmo.micro_utils.repos.cache.util
 
 import dev.inmo.micro_utils.pagination.FirstPagePagination
 import dev.inmo.micro_utils.pagination.utils.doForAllWithNextPaging
+import dev.inmo.micro_utils.pagination.utils.getAllByWithNextPaging
 import dev.inmo.micro_utils.repos.ReadCRUDRepo
 import dev.inmo.micro_utils.repos.ReadKeyValueRepo
 import dev.inmo.micro_utils.repos.ReadKeyValuesRepo
 import dev.inmo.micro_utils.repos.cache.cache.KVCache
+import dev.inmo.micro_utils.repos.pagination.getAll
 import dev.inmo.micro_utils.repos.set
 
 suspend inline fun <K, V> KVCache<K, V>.actualizeAll(
+    clear: Boolean = true,
     getAll: () -> Map<K, V>
 ) {
-    clear()
-    set(getAll())
+    set(
+        getAll().also {
+            if (clear) {
+                clear()
+            }
+        }
+    )
 }
 
 suspend inline fun <K, V> KVCache<K, V>.actualizeAll(
-    repo: ReadKeyValueRepo<K, V>
+    repo: ReadKeyValueRepo<K, V>,
+    clear: Boolean = true,
 ) {
-    clear()
-    val count = repo.count().takeIf { it < Int.MAX_VALUE } ?.toInt() ?: Int.MAX_VALUE
-    val initPagination = FirstPagePagination(count)
-    doForAllWithNextPaging(initPagination) {
-        keys(it).also {
-            set(
-                it.results.mapNotNull { k -> repo.get(k) ?.let { k to it } }
-            )
-        }
+    actualizeAll(clear) {
+        repo.getAll { keys(it) }.toMap()
     }
 }
 
 suspend inline fun <K, V> KVCache<K, List<V>>.actualizeAll(
-    repo: ReadKeyValuesRepo<K, V>
+    repo: ReadKeyValuesRepo<K, V>,
+    clear: Boolean = true,
 ) {
-    clear()
-    val count = repo.count().takeIf { it < Int.MAX_VALUE } ?.toInt() ?: Int.MAX_VALUE
-    val initPagination = FirstPagePagination(count)
-    doForAllWithNextPaging(initPagination) {
-        keys(it).also {
-            set(
-                it.results.associateWith { k -> repo.getAll(k) }
-            )
-        }
+    actualizeAll(clear) {
+        repo.getAll { keys(it) }.toMap()
     }
 }
 
 suspend inline fun <K, V> KVCache<K, V>.actualizeAll(
-    repo: ReadCRUDRepo<V, K>
+    repo: ReadCRUDRepo<V, K>,
+    clear: Boolean = true,
 ) {
-    clear()
-    val count = repo.count().takeIf { it < Int.MAX_VALUE } ?.toInt() ?: Int.MAX_VALUE
-    val initPagination = FirstPagePagination(count)
-    doForAllWithNextPaging(initPagination) {
-        keys(it).also {
-            set(
-                it.results.mapNotNull { k -> repo.getById(k) ?.let { k to it } }
-            )
-        }
+    actualizeAll(clear) {
+        repo.getAllByWithNextPaging {
+            getIdsByPagination(it)
+        }.mapNotNull { it to (repo.getById(it) ?: return@mapNotNull null) }.toMap()
     }
 }